requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.11.16
aiosignal==1.3.2
annotated-types==0.7.0
anthropic==0.51.0
anyio==4.8.0
attrs==25.3.0
bcrypt==4.3.0
beautifulsoup4==4.13.3
bidict==0.23.1
blinker==1.9.0
bs4==0.0.2
cachelib==0.13.0
cachetools==5.5.2
certifi==2025.1.31
cffi==1.17.1
chardet==5.2.0
charset-normalizer==3.4.1
click==8.1.8
cryptography==44.0.2
daily-python==0.19.1
dill==0.4.0
distro==1.9.0
dnspython==2.7.0
eventlet==0.39.1
Flask==3.1.0
Flask-Bcrypt==1.0.1
flask-cors==5.0.1
Flask-Mail==0.10.0
Flask-Session==0.8.0
Flask-SocketIO==5.5.1
frozenlist==1.5.0
git-filter-repo==2.47.0
google-ai-generativelanguage==0.6.15
google-api-core==2.24.2
google-api-python-client==2.169.0
google-auth==2.38.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.2
google-genai==1.5.0
google-generativeai==0.8.5
googleapis-common-protos==1.70.0
greenlet==3.1.1
grpcio==1.71.0
grpcio-status==1.71.0
gunicorn==23.0.0
h11==0.14.0
http_ece==1.2.1
httpcore==1.0.7
httplib2==0.22.0
httpx==0.28.1
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.5
jiter==0.9.0
Markdown==3.7
MarkupSafe==3.0.2
msal==1.32.3
msgspec==0.19.0
multidict==6.4.3
oauthlib==3.2.2
openai==1.79.0
outcome==1.3.0.post0
packaging==24.2
pdf2image==1.17.0
pdfminer.six==20250506
pdfplumber==0.11.7
pillow==11.1.0
propcache==0.3.1
proto-plus==1.26.1
protobuf==5.29.4
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
pydantic==2.10.6
pydantic_core==2.27.2
PyJWT==2.10.1
pymongo==4.11.1
PyMySQL==1.1.1
pyOpenSSL==25.0.0
pyparsing==3.2.3
PyPDF2==3.0.1
pypdfium2==4.30.1
PySocks==1.7.1
python-dotenv==1.0.1
python-engineio==4.11.2
python-socketio==5.12.1
pytz==2025.2
pywebpush==2.0.3
redis==6.2.0
requests==2.32.3
requests-oauthlib==2.0.0
rsa==4.9
selenium==4.30.0
setuptools==80.9.0
simple-websocket==1.1.0
six==1.17.0
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.6
tika==3.1.0
tqdm==4.67.1
trio==0.29.0
trio-websocket==0.12.2
typing_extensions==4.12.2
uritemplate==4.1.1
urllib3==2.3.0
webdriver-manager==4.0.2
websocket-client==1.8.0
websockets==14.2
Werkzeug==3.1.3
wsproto==1.2.0
yarl==1.19.0
